+++ admin/spellcheck.php
@@ -2,7 +2,7 @@
 /***************************************
  * http://www.program-o.com
  * PROGRAM O
- * Version: 2.6.4
+ * Version: 2.6.5
  * FILE: spellcheck.php
  * AUTHOR: Elizabeth Perreau and Dave Morton
  * DATE: 12-09-2014